% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_lengths}
\alias{add_lengths}
\title{Augment an integration analysis dataframe with region lengths.}
\usage{
add_lengths(d)
}
\arguments{
\item{d}{an integration analysis tibble}
}
\value{
a tibble, augmented to carry a length for each dimension
}
\description{
Augment an integration analysis dataframe with region lengths.
}
