% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_volume}
\alias{add_volume}
\title{Augment an integration analysis dataframe with volumes}
\usage{
add_volume(d)
}
\arguments{
\item{d}{a tibble containing lengths for each dimension (see \code{add_lengths})}
}
\value{
a tibble, augmented to carry a total volume
}
\description{
Augment an integration analysis dataframe with volumes
}
