% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{augment_raw_dataframe}
\alias{augment_raw_dataframe}
\title{Augment a raw integration analysis dataframe with lengths and volumes}
\usage{
augment_raw_dataframe(d)
}
\arguments{
\item{d}{: a raw integration dataframe}
}
\value{
a copy of the input dataframe, with dimension names canonicalized and
dimension lengths and volume added
}
\description{
Augment a raw integration analysis dataframe with lengths and volumes
}
