% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{canonicalize_dim_names}
\alias{canonicalize_dim_names}
\title{Turn dimension names into canonical form}
\usage{
canonicalize_dim_names(d)
}
\arguments{
\item{d}{an integration analysis dataframe}
}
\value{
a copy of the input dataframe, with dimension names canonicalized
}
\description{
Turn dimension names into canonical form
}
