% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dim_names}
\alias{get_dim_names}
\title{Determine the names of the column dimensions}
\usage{
get_dim_names(d)
}
\arguments{
\item{d}{an integration analysis dataframe}
}
\value{
a character vector containing the names of all dimension columns
}
\description{
Determine the names of the column dimensions
}
