% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Negated value matching.}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{vector, the values to be matched}

\item{y}{vector, the values to be matche against}
}
\value{
a vector of the same length as \code{x}
}
\description{
\%nin\% is a binary operator which returns a logical vector indicating
if there is not a match for its left operand. It is the logical negation
of \%in\%.
}
\examples{
1:10 \%nin\% 3:8
}
