% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_iteration_dataframe}
\alias{make_iteration_dataframe}
\title{Summarize an augmented raw integration dataframe by iteration}
\usage{
make_iteration_dataframe(d)
}
\arguments{
\item{d}{: an augmented raw integration dataframe}
}
\value{
a dataframe containing a per-iteration summary of the input dataframe
}
\description{
Summarize an augmented raw integration dataframe by iteration
}
