% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_dimensions}
\alias{n_dimensions}
\title{Determine the number of dimensions in an integration analysis dataframe}
\usage{
n_dimensions(d)
}
\arguments{
\item{d}{an integration analysis dataframe}
}
\value{
the dimensionality of the integrand represented
}
\description{
Determine the number of dimensions in an integration analysis dataframe
}
