% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rds_to_feather}
\alias{rds_to_feather}
\title{Convert RDS file to Feather file.}
\usage{
rds_to_feather(
  infile,
  outfile,
  compression = "uncompressed",
  compression_level = NULL,
  chunk_size = 65536L
)
}
\arguments{
\item{infile}{character, name of RDS file to be read}

\item{outfile}{character, name of Feather file to be written}

\item{compression}{see arrow::write_feather}

\item{compression_level}{see arrow::write_feather}

\item{chunk_size}{see arrow::write_feather}
}
\value{
the data read, invisibly
}
\description{
Convert RDS file to Feather file.
}
