% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_feather_or_rds}
\alias{read_feather_or_rds}
\title{Read feather or RDS data.}
\usage{
read_feather_or_rds(name)
}
\arguments{
\item{name}{The basename of the file (without the .feather or .rds extension)}
}
\value{
The object read from the file.
}
\description{
This function reads data from either a Feather or RDS file, preferring the
Feather file (if it exists).
}
