% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_finished}
\alias{select_finished}
\title{Selected finished regions from an augmented integration dataframe}
\usage{
select_finished(d)
}
\arguments{
\item{d}{: an aumented raw integration dataframe}
}
\value{
a dataframe contained only finished regions
}
\description{
Selected finished regions from an augmented integration dataframe
}
