% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_finished_by_iteration}
\alias{summarize_finished_by_iteration}
\title{Summarize finished regions by iteration}
\usage{
summarize_finished_by_iteration(d)
}
\arguments{
\item{d}{: an augmented raw integration dataframe}
}
\value{
a dataframe with one row per iteration, summarizing the input data
}
\description{
Summarize finished regions by iteration
}
